requirements.txt
Django==2.1.15
mysqlclient==2.1.0
djangorestframework
djangorestframework-simplejwt
django-filter
gunicorn
psycopg2-binary
django-dotenv
